+++ src/WebRoutes.tsx
@@ -67,7 +67,10 @@ import IndexDronerSummaryPoint from "./pages/point/droner/IndexDronerSummaryPoin
 import IndexFarmerSummary from "./pages/point/farmer/IndexFarmerSummary";
 import IndexFarmerHistorySum from "./pages/point/farmer/IndexFarmerHistorySum";
 import DetailDronerHistorySum from "./pages/point/droner/DetailDronerHistorySum";
-import MissionReport from "./pages/mission/MissionReport";
+import IndexDronerMission from "./pages/mission/droner/IndexDronerMission";
+import AddDronerMission from "./pages/mission/droner/AddDronerMission";
+import EditDronerMission from "./pages/mission/droner/EditDronerMission";
+import MissionReport from "./pages/mission/droner/MissionReport";
 
 const WebRoutes: React.FC<any> = () => {
   return (
@@ -175,8 +178,13 @@ const WebRoutes: React.FC<any> = () => {
           />
           <Route path="/MissionReport" element={<MissionReport/>} />
 
+          <Route path="/IndexDronerMission" element={<IndexDronerMission />} />
+          <Route path="/AddDronerMission" element={<AddDronerMission />} />
+          <Route
+            path="/EditDronerMission/:id"
+            element={<EditDronerMission />}
+          />
         </Route>
-
         <Route element={<PublicRoute />}>
           <Route index element={<AuthPage />} />
           <Route path="/ErrorLoginPage" element={<ErrorLoginPage />} />